+++ packages/react/src/apps/es/elements/typography/text/index.tsx
@@ -1,0 +1,39 @@
+/* ======================================================================== *
+ * Copyright 2025 HCL America Inc.                                          *
+ * Licensed under the Apache License, Version 2.0 (the "License");          *
+ * you may not use this file except in compliance with the License.         *
+ * You may obtain a copy of the License at                                  *
+ *                                                                          *
+ * http://www.apache.org/licenses/LICENSE-2.0                               *
+ *                                                                          *
+ * Unless required by applicable law or agreed to in writing, software      *
+ * distributed under the License is distributed on an "AS IS" BASIS,        *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+ * See the License for the specific language governing permissions and      *
+ * limitations under the License.                                           *
+ * ======================================================================== */
+
+/* auto generated file - do not edit */
+import { createSvgIcon, IIconAttrs, IIconContent } from '../../../../../utils';
+
+const attrs: IIconAttrs = {
+  xmlns: 'http://www.w3.org/2000/svg',
+  viewBox: '0 0 32 32',
+  width: 32,
+  height: 32,
+};
+
+const content: IIconContent[] = [
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M18,24 L18,26 L3,26 L3,24 L18,24 Z M29,18 L29,20 L3,20 L3,18 L29,18 Z M28,12 L28,14 L12,14 L12,12 L28,12 Z M9,6 L9,8 L7,8 L7,14 L5,14 L5,8 L3,8 L3,6 L9,6 Z M28,6 L28,8 L12,8 L12,6 L28,6 Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10,
+    },
+  },
+];
+
+export default createSvgIcon('text', 32, content, attrs);