+++ packages/web-component/src/apps/es/elements/typography/short-text/index.ts
@@ -1,0 +1,56 @@
+/* ======================================================================== *
+  * Copyright 2025 HCL America Inc.                                          *
+  * Licensed under the Apache License, Version 2.0 (the "License");          *
+  * you may not use this file except in compliance with the License.         *
+  * You may obtain a copy of the License at                                  *
+  *                                                                          *
+  * http://www.apache.org/licenses/LICENSE-2.0                               *
+  *                                                                          *
+  * Unless required by applicable law or agreed to in writing, software      *
+  * distributed under the License is distributed on an "AS IS" BASIS,        *
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+  * See the License for the specific language governing permissions and      *
+  * limitations under the License.                                           *
+  * ======================================================================== */
+
+/* auto generated file - do not edit */
+import { html } from 'lit';
+import { createSvgIcon, IIconAttrs, IIconContent, canDefine } from '../../../../../utils';
+import { BaseIcon } from '../../../../../utils/base-icon';
+
+const attrs: IIconAttrs = {
+  xmlns: 'http://www.w3.org/2000/svg',
+  viewBox: '0 0 24 24',
+  width: 24,
+  height: 24
+};
+
+const content: IIconContent[] = [
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M13.5,18 L13.5,19.5 L2.25,19.5 L2.25,18 L13.5,18 Z M21.75,13.5 L21.75,15 L2.25,15 L2.25,13.5 L21.75,13.5 Z M6.75,4.5 L6.75,6 L5.25,6 L5.25,10.5 L3.75,10.5 L3.75,6 L2.25,6 L2.25,4.5 L6.75,4.5 Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  }
+];
+
+export const ICON_NAME = 'icon-short-text';
+export class WebComponentIcon extends BaseIcon {
+  render() {
+    return html`${createSvgIcon(content, attrs)}`;
+  }
+}
+
+if (canDefine && !customElements.get(ICON_NAME)) {
+  customElements.define(ICON_NAME, WebComponentIcon);
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    [ICON_NAME]: WebComponentIcon;
+  }
+}