+++ packages/web-component/src/apps/es/layers--alt/index.ts
@@ -1,0 +1,86 @@
+/* ======================================================================== *
+  * Copyright 2025 HCL America Inc.                                          *
+  * Licensed under the Apache License, Version 2.0 (the "License");          *
+  * you may not use this file except in compliance with the License.         *
+  * You may obtain a copy of the License at                                  *
+  *                                                                          *
+  * http://www.apache.org/licenses/LICENSE-2.0                               *
+  *                                                                          *
+  * Unless required by applicable law or agreed to in writing, software      *
+  * distributed under the License is distributed on an "AS IS" BASIS,        *
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+  * See the License for the specific language governing permissions and      *
+  * limitations under the License.                                           *
+  * ======================================================================== */
+
+/* auto generated file - do not edit */
+import { html } from 'lit';
+import { createSvgIcon, IIconAttrs, IIconContent, canDefine } from '../../../utils';
+import { BaseIcon } from '../../../utils/base-icon';
+
+const attrs: IIconAttrs = {
+  xmlns: 'http://www.w3.org/2000/svg',
+  viewBox: '0 0 24 24',
+  width: 24,
+  height: 24
+};
+
+const content: IIconContent[] = [
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M7 18H21V19.5H7V18Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  },
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M7 13.5H21V15H7V13.5Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  },
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M7 9H21V10.5H7V9Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  },
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M3 4.5H17V6H3V4.5Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  }
+];
+
+export const ICON_NAME = 'icon-layers-alt';
+export class WebComponentIcon extends BaseIcon {
+  render() {
+    return html`${createSvgIcon(content, attrs)}`;
+  }
+}
+
+if (canDefine && !customElements.get(ICON_NAME)) {
+  customElements.define(ICON_NAME, WebComponentIcon);
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    [ICON_NAME]: WebComponentIcon;
+  }
+}