+++ packages/web-component/src/apps/es/page--elements--h4/index.ts
@@ -1,0 +1,68 @@
+/* ======================================================================== *
+  * Copyright 2025 HCL America Inc.                                          *
+  * Licensed under the Apache License, Version 2.0 (the "License");          *
+  * you may not use this file except in compliance with the License.         *
+  * You may obtain a copy of the License at                                  *
+  *                                                                          *
+  * http://www.apache.org/licenses/LICENSE-2.0                               *
+  *                                                                          *
+  * Unless required by applicable law or agreed to in writing, software      *
+  * distributed under the License is distributed on an "AS IS" BASIS,        *
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+  * See the License for the specific language governing permissions and      *
+  * limitations under the License.                                           *
+  * ======================================================================== */
+
+/* auto generated file - do not edit */
+import { html } from 'lit';
+import { createSvgIcon, IIconAttrs, IIconContent, canDefine } from '../../../utils';
+import { BaseIcon } from '../../../utils/base-icon';
+
+const attrs: IIconAttrs = {
+  xmlns: 'http://www.w3.org/2000/svg',
+  viewBox: '0 0 24 24',
+  width: 24,
+  height: 24
+};
+
+const content: IIconContent[] = [
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M18.75 6.75H17.25V12.75H15V6.75H13.5V14.25H17.25V17.25H18.75V14.25H19.5V12.75H18.75V6.75Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10,
+      fillRule: 'evenodd',
+      clipRule: 'evenodd'
+    }
+  },
+  {
+    elem: 'path',
+    attrs: {
+      d: 'M9 6.75V11.25H6V6.75H4.5V17.25H6V12.75H9V17.25H10.5V6.75H9Z',
+      fill: 'currentColor',
+      stroke: 'none',
+      strokeLinejoin: 'round',
+      strokeMiterlimit: 10
+    }
+  }
+];
+
+export const ICON_NAME = 'icon-page-elements-h4';
+export class WebComponentIcon extends BaseIcon {
+  render() {
+    return html`${createSvgIcon(content, attrs)}`;
+  }
+}
+
+if (canDefine && !customElements.get(ICON_NAME)) {
+  customElements.define(ICON_NAME, WebComponentIcon);
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    [ICON_NAME]: WebComponentIcon;
+  }
+}